+++ classes/local/checks/plugins_restore.php
@@ -21,6 +21,7 @@ use moodle_url;
 use tool_vault\api;
 use tool_vault\constants;
 use tool_vault\local\helpers\siteinfo;
+use tool_vault\local\models\dryrun_model;
 
 /**
  * Check plugins version on restore
@@ -35,6 +36,21 @@ class plugins_restore extends check_base_restore {
      * Evaluate check and store results in model details
      */
     public function perform(): void {
+        global $CFG;
+        /** @var dryrun_model $parent */
+        $parent = $this->get_parent();
+        $backupbranch = $parent->get_metadata()['branch'];
+        if ($backupbranch > $CFG->branch) {
+            // Skip this check if backup has a higher major moodle version, or it will be unreadable
+            // and full of confusing errors.
+            $this->model->set_details([
+                'list' => [],
+                'standardplugins' => [],
+                'skipped' => true,
+            ])->save();
+            return;
+        }
+
         $excludedplugins = ['tool_vault']; // TODO more from settings.
         $pluginlist = array_diff_key(siteinfo::get_plugins_list_full(true), array_fill_keys($excludedplugins, true));
         $parent = $this->get_parent();
@@ -46,6 +62,8 @@ class plugins_restore extends check_base_restore {
         $this->model->set_details([
             'list' => $list,
             'standardplugins' => $standardplugins,
+            'backupbranch' => $backupbranch,
+            'backupkey' => $parent->backupkey,
         ])->save();
     }
 
@@ -130,10 +148,24 @@ class plugins_restore extends check_base_restore {
      * @return array
      */
     protected function missing_plugins(bool $includestandard = true): array {
-        return array_filter($this->model->get_details()['list'], function($info, $pluginname) use ($includestandard)  {
+        $plugins = array_filter($this->model->get_details()['list'], function($info, $pluginname) use ($includestandard)  {
             return empty($info[1]) && !empty($info[0])
                 && ($includestandard || !$this->is_standard_plugin($pluginname));
         }, ARRAY_FILTER_USE_BOTH);
+
+        // Do not return subplugins of missing plugins as individual entries.
+        foreach (array_keys($plugins) as $pluginname) {
+            if (!empty($plugins[$pluginname][2]['parentismissing'])) {
+                $parent = $plugins[$pluginname][0]['parent'];
+                if (array_key_exists($parent, $plugins)) {
+                    $plugins[$parent][0]['subplugins'] = $plugins[$parent][0]['subplugins'] ?? [];
+                    $plugins[$parent][0]['subplugins'][$pluginname] = $plugins[$pluginname][0];
+                    unset($plugins[$pluginname]);
+                }
+            }
+        }
+
+        return $plugins;
     }
 
     /**
@@ -143,6 +175,9 @@ class plugins_restore extends check_base_restore {
      */
     public function success(): bool {
         if ($this->model->status !== constants::STATUS_FINISHED) {
+            return false;
+        }
+        if (!empty($this->model->get_details()['skipped'])) {
             return false;
         }
         if ($this->problem_plugins()) {
@@ -170,7 +205,9 @@ class plugins_restore extends check_base_restore {
                 return get_string('addonplugins_success', 'tool_vault');
             }
         } else {
-            if ($this->problem_plugins()) {
+            if (!empty($this->get_model()->get_details()['skipped'])) {
+                return get_string('addonplugins_fail_skipped', 'tool_vault');
+            } else if ($this->problem_plugins()) {
                 return get_string('addonplugins_fail', 'tool_vault');
             } else {
                 return get_string('addonplugins_fail_missing', 'tool_vault');